Fixed 'getResourceProviderId' for operations without resources.
The code in 'getResourceProviderId' was written in a way assuming that
any operation containing a list resources always contained at least a
single resource. This is both incorrect since such operations pass
validation, and also neglects to check a crucial precondition since
accessing a non-existent element in this context would lead to a check
failure in protobuf and a subsequent abort.

This code updates 'getResourceProviderId' to check the precondition of
operations holding at least one element.

Review: https://reviews.apache.org/r/64425