Added plumbing for master to reconcile offer operations with agent.
This patch adds the RECONCILE_OFFER_OPERATIONS event to the resource
provider API, along with the internal message
'ReconcileOfferOperationsMessage' used for explicit operation
reconciliation between master and agent. Handlers for these are
added to the agent and resource provider manager as well.

This explicit reconciliation is useful in cases where an agent's
'UpdateSlaveMessage' races with an incoming task launch so that
the master's view of the agent's state is not consistent with the
agent's actual state when the 'UpdateSlaveMessage' was sent.

Review: https://reviews.apache.org/r/63804/