Fixed bug where we didn't check non-localhost URIs were absolute.
The previous change here subtley missed that we check if the path is
absolute if it is a URI, based on `fileUri`, and stopped setting
`fileUri = true` when the path was a `file://` URI, but not
`file://localhost`.

Now we set `fileUri = true` if the URI starts with `file://`, and then
use `path::from_uri()` to strip `file://`, and `strings::remove()` to
strip the remaining `localhost` prefix.