Added flag to pass VolumeProfileAdaptor module to SLRP.
This adds an agent flag to choose the VolumeProfileAdaptor module
to load and passes this information down to the storage resource
providers. A single module can potentially be used by multiple
storage resource providers.

We choose to pass the module along as a global variable because the
module is not needed by many of the components in the injection path.
Specifically, the Storage Local Resource Provider will need the module,
but its parents, the Local Resource Provider and Resource Provider
Daemons will not.

Review: https://reviews.apache.org/r/64612