Windows: Fixed type in job object `pid_list` struct.
This type was erroneously declared as `ULONG_PTR` because the
documentation states it as such. This is true if the array is
dynamically allocated, then the struct ends with a pointer to the first
element of the array. However, we avoid this allocation by statically
defining the struct with an array the size of `max_pids` for process
identifiers. PIDs on Windows are `DWORD` typed, and since the array is
defined statically, the type is the element type, not the pointer type.

The compiler warning for this was missed in the noise of existent
compiler warnings on Windows.