Fixed a bug in handling standalone container recovery.
It's possible that the pid is not known for a standalone container. For
instance, the agent crashes after the containerizer checkpoints the
runtime directory, but before it is able to checkpoint the pid. In that
case, we just assume that the child will exit due to the broken control
pipe and we can move to destroy without recovering it.

Review: https://reviews.apache.org/r/64623