Added standalone isolator capability.
Recently, we introduced the standalone container suport (MESOS-7302).
Some isolators might not be able to handle standalone containers.
Therefore, we introduce an isolator capability (similar to nesting
capability) so that we won't try to invoke those isolators that do not
support standalone containers when dealing with a standalone container.

Review: https://reviews.apache.org/r/64624