Windows: Removed `/debug:fastlink` linker flag.
With Visual Studio 2017, `fastlink` is now the default, so this is
superfluous. Morever, this particular CMake variable also passes the
flag to `lib.exe` when generating a static library, and this emits a
warning because `lib.exe` does not accept this flag. It was intended to
be sent to the `link.exe` linker, and so should not have included
`STATIC` in the list of types.