Windows: Patched protobuf to fix warnings.
The patch was generated by diffing the following commit against tag
v3.5.0, this was merged upstream in
https://github.com/google/protobuf/pull/4000:

commit 80809271a
Author: Rodrigo Hernandez <kwizatz@aeongames.com>
Date:   Mon Dec 4 19:04:42 2017 -0600

    Using binary one's complement to negate an unsigned int

    This removes a Visual Studio warning:

    warning C4146: unary minus operator applied to unsigned type, result
    still unsigned.

And the addition of three static casts to make the return of
`std::distance` match the LHS type.

With the patch applied, we can stop disabling all warnings for protobuf
code (which didn't entirely work anyway, because these warnings were in
public headers included everywhere in Mesos).