Fixed a bug related to offer operation application in the agent.
The bug is introduced in this patch:
https://reviews.apache.org/r/64477/

Given that we also keep track of each resource provider's total
resources in addition to the total resources of the agent, we need to
make sure we update those totals after applying an operation.

The bug may manifest as a CHECK failure in the agent that checks if
`totalResources` of the agent is a super set of all the resource
provider resources.

Review: https://reviews.apache.org/r/64589/