Windows: Updated cURL to version 7.57.0.
The cURL source and build have improved significantly since we pulled in
the previous versoin (7.43.0). Most notably, the new version now
supports SChannel (native Windows crypto) on Windows, so we no longer
need to use OpenSSL. This benefits us two-fold: it fixes the issue where
OpenSSL on Windows does not come with a certificate bundle (and thus
caused any downloads over TLS to fail because the certificates couldn't
be checked), and means we can enable TLS support for anything that uses
cURL unconditionally, as it "just works".

Because we set `-DCMAKE_USE_WINSSL=ON` uncondtionally, we had to the
`crypt32` Windows library to the interface link libraries for `libcurl`.

Between the old and new versions, cURL changed some aspects of their
build system:

1. The Debug configuration now calls the executable `curl-d.exe`, so we
   had to updated its imported location and explicitly rename it to
   `curl.exe` in the copy step.

2. We continued to disable building the tests, but the option changed
   its name from `BUILD_CURL_TESTS` to `BUILD_TESTING`. This is now
   especially important because building the tests require Perl, which
   we definitely do not want to add.

3. We disabled LDAP because it didn't build without further
   setup (OpenLDAP), and we don't currently need it.

Review: https://reviews.apache.org/r/64720/