Fixed an agent assertion.
Initially it was valid to assert that 'checkpointResources' would
never be called for an agent with resource providers, but we have
sinced moved to an implementation where we might perform checkpointing
of agent resources (but never local resource provider resources).

This patch adjusts an assertion in the agent's 'checkpointResources'
function to that reality.

Review: https://reviews.apache.org/r/64750/