Renamed offer operation to operation.
The motivation for this renaming is that the Operation workflow
(including status update, retry, etc.) might be re-used for non offer
related operations in the future. As a result, we intentionally make the
name more general to capture potential future use cases.

This patch also made the naming of protobuf message types, fields and
variable names that are related to Operations more consistent.

This patch also renamed master registry operation to `RegistryOperation`
to distinguish that from `Operation`.

Review: https://reviews.apache.org/r/64762/