Updated support/mesos-style.py to build the virtualenv less often.
Instead of just checking if the virtual environment exists, we now
also check if it is required to lint the modified files or not.

The virtual environment will thus be built when using the script as
a git hook only if JavaScript or Python files have been modified.

Review: https://reviews.apache.org/r/64683/