Cleaned up ContainerInfo generation logic in the slave.
After this patch, the invariant becomes that for command tasks, we
always set the ContainerInfo of the generated ExecutorInfo to be the
same as that in TaskInfo (if exists). This simplifies the logic when we
actually generate ContainerInfo for containerizer during launching
phase.

This also made the logic introduced in the following patch more readable
and maintainable: https://reviews.apache.org/r/63598

Review: https://reviews.apache.org/r/64811