Fixed resource provider driver disconnection handling.
The expectation for disconnection of the resource provider driver is
that disconnection handlers of the resource provider would be invoked
and a new connection would be detected.

This patch fixed the issue by transition the future returned by
ConstantEndpointDetector into DISCARDED if a "discard" is initiated by
the caller. This will properly trigger `detected` callback to be called.

This patch is based on: https://reviews.apache.org/r/64806/

Review: https://reviews.apache.org/r/64856