Fixed a quota headroom violation during quota allocation.
In the quota role allocation stage, if a role gets some resources
on an agent to meet its quota, it will also get all other resources
on the same agent that it does not have quota for. This may violate
the guarantees of other roles that have set quota on these resources.

We fix this issue during the quota allocation stage by ensuring that,
if a role has no quota set for a scalar resource, it will get that
resource only when two conditions are both met:

  (1) It is tentatively being allocated some other resources on the
      same agent to meet its quota, or some reservations (which may
      or may not involve quota resources); and

  (2) After allocating those resources, quota headroom is still
      above the required amount.

Also refactored the fine-grained quota allocation logic.

Review: https://reviews.apache.org/r/64793/