Added a test for quota allocation including non-quota resources.
This test verifies the behavior of allocating a resource to quota
roles that have no quota set for that particular resource (e.g.
allocating memory to a role with only quota set for CPU). If a
role has no quota set for a resource, it will get that resource
only when two conditions are both met:

  (1) It is tentatively being allocated some other resources on the
      same agent to meet its quota, or some reservations (which may
      or may not involve quota resources); and

  (2) After allocating those resources, quota headroom is still
      above the required amount.

Review: https://reviews.apache.org/r/64794/