Made sure continuations using actor state are executed in actor context.
In `e273efe6976434858edb85bbcf367a02e963a467` we introduced layer
checkpointing to the provisioner, but did not make sure that
continutations making use of internal state of the provisioner actor
installed on futures where always executed in the actor's context.
This was problematic as continuations could be executed while actor
state was changing (data races), or after the the actor had terminated
(use after free).

In this patch we instead defer execution of continutions to the actor
owning the data.

This issue was identified with the clang-tidy `mesos-this-capture`
check.

Review: https://reviews.apache.org/r/64885/