Abort libprocess when a Process throws an uncaught exception.
Previously, we logged the exception, terminated the throwing
Process, and continued to run. However, currently there exists
no known user-level code that I'm aware of that handles the
unexpected termination due to an uncaught exception.

Generally, this means that when an exception is thrown (e.g.
a bad call to std::map::at), the process terminates with a log
message but things get "stuck" and the user has to debug what
is wrong / kill the process.

Libprocess would likely need to provide some primitives to
better support handling unexpected termination of a Process
in order for us to provide a strategy where we continue running.

Review: https://reviews.apache.org/r/64939