Future-proofed use of agent capabilities in tests.
Even though currently the resource provider capability is the only
capability which can be toggled by users, when examining agent flags
we expect either no capabilities at all or a full set of capabilities
(including both toggleable and required capabilities).

This patch cleans up test code agent capabilities by delegating the
bulk of the work to a helper function providing a set of
default-enabled capabilities and only adds a single capability to that
set. This not only makes it clearer which exact capability a test
cares about, but also future-proofs the code for the case where we
extend the set of required capabilities in the future.

Review: https://reviews.apache.org/r/64891/