Fixed RecoverTest.CatchupTruncated test flakiness.
Most likely the "lock already held by process" LevelDB error was caused
by a Shared<Replica> still retained by one of the processes when the
test tries to recreate it. This change makes sure that the test code is
the only owner of the replica.

Review: https://reviews.apache.org/r/64938/