Removed the inaccurate `slaves.transitioning` function in the master.
This function used to return whether the agent was transitioning
between states, however it was updated in e510813f93e253480005ce
to return only when the agent is recovered.

This patch removes the now confusing function in favor of just
directly checking the state we care about in the reconciliation
logic. Since there were no other usages, the function is removed.

Review: https://reviews.apache.org/r/64968