Fixed the order of parameters of `is_specialization_of` in stout.
Given `std::is_base_of<T, U>`, the interpretation of the question is:
"is `T` base of `U`?". Similarly, in this case want to ask the question
"is some type a specialization of a template?" As such, the order of
template paramters should be `is_specialization_of<Type, Template>`.

Review: https://reviews.apache.org/r/65054