Replaced `ResourcesState::recoverResources` with `state::read`.
We have 2 versions of `protobuf::(write|read)`. One that operates on
a file descriptor, and another that works with file paths. The file
descriptor version is used to open a file in append mode and appends
to the file as events (e.g., status updates) arrive. We then read
the messages back one at a time until we eventually read nothing,
at which point `protobuf::read` returns a `None` indicating EOF.
The path version on the other hand simply tries to read the full
expected state, and returns an error if it's not able to.

The existing `ResourcesState::recoverResources` used the file descriptor
version of `protobuf::read`, reading one message at a time as described
above. However, the resources being read back is written using
`state::checkpoint`. `state::checkpoint` uses the path version of
`protobuf::read`, and also provides atomic writes. We therefore do not
need to deal with partial writes, and expect to read the full state.
As such, we can now simply use `state::read` to read the resources back.

Review: https://reviews.apache.org/r/65026