Added some defensive checks for persistent volume sharing check.
This patch adds some defensive checks for detecting if a persistent
volume is shared by multiple containers. It's currently not a bug
because of a `contains` check above in the loop body and we set
`info->resources` only after we update the volumes. But this is a bit
fragile and should be avoided.

See more details in MESOS-8356.

Review: https://reviews.apache.org/r/65050/