Fixed post-reviews for posting reviews from `master` branch.
The post-reviews script allows to specify the tracking branch against
which to post reviews; we default to creating patches against
`master`.

The script also contains some protection against posting reviews from
the tracking branch, but this protection seems to predate the tracking
branch configuration flag. Instead we also disallow posting patches
from the `master` branch, even if the tracking branch is not `master`.

This patch fixes this sanity check to check that the current and
tracking branch are different.

Review: https://reviews.apache.org/r/65073/