Fixed a potential race condition in the test infrastructure.
There was a race condition leading to flaky
`LinuxCapabilitiesIsolatorFlagsTest.ROOT_IsolatorFlags` test.
This test launches successively multiple agents, while reusing the
same variable. After reassigning the value of the variable, agent's
d-tor is called. If agent recovery is not yet completed, then some
orphaned container might blink in the agent's d-tor as it is
described in the comment to the code.

Review: https://reviews.apache.org/r/64770/