Ported `os::which` to Windows.
Because `os::which` still lived in `posix/os.hpp`, it was refactored
into its own `os/which.hpp` (which the respective `os/posix/which.hpp`
and `os/windows/which.hpp`. Consumers of this will need additionally
include the new header, instead of just `os.hpp`.

The differences in implementation from POSIX to Windows are:

* Split the `PATH` on `;` not `:`.
* Also loop over `PATHEXT` because executables on Windows end with one
  of a set of extensions.
* Removed the permissions check because it is not applicable on
  Windows (instead, an executable ends with an extension in `PATHEXT`,
  see above).

Review: https://reviews.apache.org/r/65144