Windows: Fixed docker executor `PATH` variable.
The `docker` executable is not usually installed in
`os::host_default_path()` on Windows, so the Executor cannot find it.
Now, before launching the Executor, the Agent finds the directory
containing `docker` and prepends it to the `PATH` given to the Executor
so that both the Executor and Agent use the same `docker`.

Review: https://reviews.apache.org/r/65147