Stopped applying empty executor resources in `network/ports` recovery.
At recovery time, the containerizer sends only the executor resources,
but ports are typically allocated against tasks. This means that we
don't know about the complete set of container resources until the
containerizer makes the subsequent `update()` call. If we perform a
container ports check between the two calls, we can erroneously kill
the container because we believe it to have no allocated ports.

Review: https://reviews.apache.org/r/65153/