Narrowed task sandbox permissions from 0755 to 0750.
Since task sandboxes can contain private data, we should not
make them accessible to others by default. This changes all the
places that create a task sandbox directory to use a helper API
`slave::paths::createSandboxDirectory` that consistently deals
with setting the directory mode and ownership.

A number of tests depended on the previous behavior where
failing to change the ownership was logged but did not cause
a failure. Depending on the test, these were updated to either
disable the agent `switch_user` flag, or to specify the current
user in the task launch message.

Review: https://reviews.apache.org/r/64630/