Fixed a libprocess finalization deadlock due to Queue discard support.
Once discard support for Queue::get was introduced, this makes it
possible for the finalization logic in libprocess to try to acquire
the `socket_mutex` recursively:

  (1) process::finalize acquires socket_mutex, and discards __s__.
  (2) The get from LibeventSSLSocket::accept_queue gets discarded.
  (3) This then executes the process.cpp logic in on_accept, which
      will always re-aquire the `socket_mutex`, even though the
      accepted socket came out as discarded.

The fix applied here is to also stop the accept loop when the socket
was discarded. This avoids acquiring the `socket_mutex` in the
discarded case.

Also, this avoids logging a failure message in the discarded case,
and instead consistently logs when the accept loop is stopped.

Also, wrote a comment within `LibeventSSLSocket::accept` that
explains a potential issue due to MESOS-8448.

Review: https://reviews.apache.org/r/65183