Fixed flaky EOFBeforeRecv test.
This patch is a small correction of commit d4185d4a11.
Before this patch, if EOF event for a ssl socket is received before
a client called `recv`, then this EOF event is lost. This leads to
hanging `recv` on the ssl socket after calling `shutdown`.

Review: https://reviews.apache.org/r/65179/