Windows: Tied task lifetimes to executors.
To enable recovery of checkpointed tasks, the agent must be able to die
without also killing the executors and tasks, thus we cannot set the
"job object kill on close" limit unconditionally. However, the executors
must still be able to kill their tasks when they die, so we explicitly
enable this limit through a parent hook when launching the container for
the task. In this way, the agent can be restarted (e.g. for an upgrade)
without killing the executors, but the executors are still capable of
killing their tasks on catastrophic death.

Review: https://reviews.apache.org/r/65400