Implemented `net::socket()` for Windows using `WSASocket`.
Rather than reusing the existing `net::socket()` wrapper for Windows,
the original wrapper was moved to `posix/socket.hpp`, and a new wrapper
was written in `windows/socket.hpp` which uses `::WSASocket()` instead
of `::socket()` and takes an optional set of Windows socket flags.

We enable `WSA_FLAG_OVERLAPPED` so that the socket supports overlapped
I/O operations (the default behavior of `::socket()`; MSDN states that
most sockets should be created with this flag set, so it is a reasonable
default.

Furthermore, we enable `WSA_FLAG_NO_HANDLE_INHERIT`, which disables
inheritance of socket handles. This is set by default because it is not
possible to change afterwards, we never use inheritable sockets, and we
do not want children to inherit our sockets. This fixes bugs due to
child processes inadvertantly inheriting sockets. For example, in the
recovery process of the Mesos agent, executors were unable to
re-register because of leaked sockets, preventing communication to the
new agent.

Review: https://reviews.apache.org/r/65405