Windows: Used Unicode API to duplicate sockets.
This code was still using the macro `::WSADuplicateSocket` instead of
the explicit Unicode version `::WSADuplicateSocketW`. Because `UNICODE`
has been defined, the code logically has not changed, but the comment
was out-of-date, and the warning no longer needs to be disabled.

We've left file handle duplication alone for a subsequent patch to fix,
as there are side effects from the implicit inheritance semantics of
`::_dup` that we currently rely on.

Review: https://reviews.apache.org/r/65402