Fixed a race in the test `ROOT_MultiTaskgroupSharePidNamespace`.
In the test `DefaultExecutorTest.ROOT_MultiTaskgroupSharePidNamespace`,
we read the file `ns` in each of the two task's sandbox and check if
their contents (the pid namespace of the task itself) are same. However
it is possible we do the read for the second task after that file is
created but before it is written, i.e., the content we read from the
`ns` file of the second task would be empty which will cause the check
failed.

In this patch, we read the file `ns` for each task in a while loop, and
only break from the loop when both task's files are not empty.

Review: https://reviews.apache.org/r/65278