Added os::eraseenv to clear the old environment value.
On Linux, updating the environment with unsetenv(3) does
not change the value that other processes can inspect
(e.g. through `/proc/$pid/environ`). We have to overwrite
the existing value (if any) prior to calling unsetenv(3)
if we want to ensure that the value is really gone.

Review: https://reviews.apache.org/r/63422/