Fixed a libprocess crash when tests clear the authorization callbacks.
The authorization callbacks are not mutated in a thread-safe manner:
a `Process` may be calling a callback while the callbacks are being
deleted by another thread.

The fix here for now is to make the mutation and access synchronized
under a mutex. The mutex is non-recursive, which means that the
callbacks must not be cleared or re-set from within a callback. This
is already a requirement, since a callback calling in to clear or
re-set the callbacks is thereby deleting itself!

Some TODOs were left for performance improvements. Since the HTTP
likely has much more significant bottlenecks, I left these as TODOs.

Review: https://reviews.apache.org/r/65305